Refact with seperated full/light client
