+++ node/service/Cargo.toml
@@ -108,7 +108,6 @@ with-bifrost-runtime = [
 	"bifrost-runtime",
 ]
 with-bifrost-polkadot-runtime = [
-	"bifrost-runtime",
 	"bifrost-polkadot-runtime",
 ]
 with-all-runtime = [
